#137: Move package digest varification out of the proofdb. This function reads proofs from the proofdb but shouldn't be a function of the proofdb.
