+++ crev-lib/src/lib.rs
@@ -10,18 +10,25 @@ pub mod repo;
 pub mod staging;
 pub mod util;
 
+pub use self::local::Local;
+pub use crate::proofdb::{ProofDB, TrustDistanceParams};
 use crate::{prelude::*, proofdb::TrustSet};
-use crev_data::Digest;
+pub use activity::{ReviewActivity, ReviewMode};
+use crev_data::{
+    self,
+    proof::{
+        review::{self, Rating},
+        trust::TrustLevel,
+        CommonOps,
+    },
+    Digest, Id,
+};
 use failure::format_err;
 use std::{
-    collections::HashSet,
+    collections::{HashMap, HashSet},
     fmt,
     path::{Path, PathBuf},
 };
-
-pub use self::local::Local;
-pub use crate::proofdb::{ProofDB, TrustDistanceParams};
-pub use activity::{ReviewActivity, ReviewMode};
 
 /// Trait representing a place that can keep proofs (all reviews and trust proofs)
 ///
@@ -128,6 +135,48 @@ impl fmt::Display for VerificationStatus {
     }
 }
 
+pub fn verify_package_digest(
+    digest: &Digest,
+    trust_set: &TrustSet,
+    requirements: &VerificationRequirements,
+    db: &proofdb::ProofDB,
+) -> VerificationStatus {
+    let reviews: HashMap<Id, review::Package> = db
+        .get_package_reviews_by_digest(digest)
+        .map(|review| (review.from().id.clone(), review))
+        .collect();
+    // Faster somehow maybe?
+    let reviews_by: HashSet<Id, _> = reviews.keys().cloned().collect();
+    let trusted_ids: HashSet<_> = trust_set.trusted_ids().cloned().collect();
+    let matching_reviewers = trusted_ids.intersection(&reviews_by);
+    let mut trust_count = 0;
+    let mut negative_count = 0;
+    for matching_reviewer in matching_reviewers {
+        let review = &reviews[matching_reviewer].review;
+        if !review.is_none()
+            && Rating::Neutral <= review.rating
+            && requirements.thoroughness <= review.thoroughness
+            && requirements.understanding <= review.understanding
+        {
+            if TrustLevel::from(requirements.trust_level)
+                <= trust_set.get_effective_trust_level(matching_reviewer)
+            {
+                trust_count += 1;
+            }
+        } else if review.rating <= Rating::Negative {
+            negative_count += 1;
+        }
+    }
+
+    if negative_count > 0 {
+        VerificationStatus::Negative
+    } else if trust_count >= requirements.redundancy {
+        VerificationStatus::Verified
+    } else {
+        VerificationStatus::Insufficient
+    }
+}
+
 /// Check whether code at this path has reviews, and the reviews meet the requirements
 pub fn dir_or_git_repo_verify(
     path: &Path,
@@ -142,7 +191,12 @@ pub fn dir_or_git_repo_verify(
         Digest::from_vec(util::get_recursive_digest_for_dir(path, ignore_list)?)
     };
 
-    Ok(db.verify_package_digest(&digest, trusted_set, requirements))
+    Ok(verify_package_digest(
+        &digest,
+        trusted_set,
+        requirements,
+        &db,
+    ))
 }
 
 /// Check whether code at this path has reviews, and the reviews meet the requirements
@@ -156,7 +210,12 @@ pub fn dir_verify(
     requirements: &VerificationRequirements,
 ) -> Result<crate::VerificationStatus> {
     let digest = Digest::from_vec(util::get_recursive_digest_for_dir(path, ignore_list)?);
-    Ok(db.verify_package_digest(&digest, trusted_set, requirements))
+    Ok(verify_package_digest(
+        &digest,
+        trusted_set,
+        requirements,
+        &db,
+    ))
 }
 
 pub fn get_dir_digest(path: &Path, ignore_list: &fnv::FnvHashSet<PathBuf>) -> Result<Digest> {
